delete timesheets

+++ src/server/timesheet/validation.ts
@@ -2,8 +2,8 @@ import { z } from "zod";
 
 export const timesheetSchema = z.object({
   date: z.string(),
-  hours: z.coerce.number().min(0).max(24),
-  minutes: z.coerce.number().min(0).max(60),
+  hours: z.coerce.number().int().min(0).max(24),
+  minutes: z.coerce.number().int().min(0).max(60),
   description: z?.string().max(548),
   id: z.number().optional(),
 });
